+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added basic line history
 
+- Added EditorBuilder for more ergonomic construction of editors
+
 ## [0.1.0] - 2022-03-14
 
 - Initial release